a2a/hotel/pyproject.toml
[project]
name = "hotel"
version = "0.1.0"
description = "Hotel booking agent with A2A server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "strands-agents[a2a]",
    "bedrock-agentcore",
    "openai",
]
